Do not log ErrSmallBuffer for request headers (#3)
unless LogAllErrors is set. fasthttp returns a 431 error when the request
headers are too large. Most other HTTP servers do not log an error when
the request header exceeds the limit. When serving HTTP requests
directly to browsers, it is normal to occasionally hit the limit due to
broken browsers or broken sites linking to your resource. These issues
are generally outside of your control in the same way the already
ignored errors are.

Note that ErrSmallBuffer _should_ always be logged for responses as this
should always be under your control.